requirements.txt
gensim==4.3.2
ipykernel==6.25.1
keras==2.13.1
matplotlib==3.7.2
mediapipe==0.10.3
openai-whisper==20230314
opencv-python==4.8.0.76
pandas==2.0.3
plotly==5.16.1
PyYAML==6.0.1
requests==2.31.0
tensorflow==2.13.0
torch==2.0.1
tqdm==4.66.1
transformers==4.32.1
SpeechRecognition==3.10.0
scipy==1.11.2
numpy==1.24.3
flask==2.3.3
torchvision
protobuf==1.20.3
grpcio-status=1.41.0